kPsychTripleBufferWorkaround
See ConservVRAMSettings.m for what it does.
In hindsight, this workaround is probably not worth the added complexity. In comparison to kPsychBusyWaitForVBLBeforeBufferSwapRequest, kPsychTripleBufferWorkaround cuts some corners here and there (maybe most notably not waiting for 250µs when timing is tight already), but the major issue turned out to be a glClear() inefficiency, which hopefully will be fixed in some of the next INTEL UHD graphics driver updates.
One should be aware that PTB cannot determine when the SwapBuffers() requests are finally served. But comparing the timestamps to photodiode measurements, while systematically varying the time of calling SwapBuffers() within the refresh cycle, suggests that the deadline for calling SwapBuffers() just in time for still being served at VBL N lies safely before VBL N (and far away from VBL N-1), meaning that the SwapBuffers() requests will be reliably deferred to the next VBL unless the timestamps indicate otherwise, even more so when the timestamps are beam-position corrected (the default). If timestamps indicate otherwise, the according flips could still have been executed in time, meaning we still can have false detection of frame timing errors, but there will be no misses of timing errors.