Screen-TransformTexture: Make source texture dispensable after use
Problem: Releasing a texture after having used it as a source texture with Screen('TransformTexture',...) can crash Matlab without any further message. This problem occurs when having left it to PTB to create the target texture (by passing [] to the 'targetTexture' parameter when calling Screen('TransformTexture',...)), because PTB creates the target texture as a child of the source texture and, therefore, requires the source texture to remain intact as long as the target texture is being used. Otherwise, i.e. if releasing the source texture before using the target texture, internal functions like PsychGetParentWindow(target texture) will use an invalid reference and things go south if the referenced memory content has changed meanwhile (which surprisingly often is not the case, actually).
Solution: Make the target texture a sibling of the source texture rather than a child.